backend/requirements.txt
fastapi==0.109.0
uvicorn==0.25.0
python-dotenv==1.0.0
google-auth==2.26.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.106.0
gspread==5.12.4
requests==2.31.0
pydantic==2.8.0
python-multipart==0.0.6
